moved initcontainer e2e tests to a separate file
